+++ changelog.md
@@ -6,8 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
-- Improve Publish.WriteTo performance
-- Improve Connect.WriteTo performance
+- Optimise PubAck, ConnAck, Publish, Connect
 
 ## [0.12.0] 2022-09-03